### 2019-10-14 #1807 @whedon
[REVIEW]: momepy: Urban Morphology Measuring Toolkit
**Submitting author:** @martinfleis (<a href="http://orcid.org/0000-0003-3319-3366">Martin Fleischmann</a>)
**Repository:** <a href="https://github.com/martinfleis/momepy" target ="_blank">https://github.com/martinfleis/momepy</a>
**Version:** v0.1.0
