### 2019-10-17 899d205
BUG: fix for networkx2.4 (#96)
