### 2019-10-17 808d112
BUG: StreetProfile overwrites distance (#97)
