### 2019-10-18 b791f3c
BUG: encoding issue when installong from source on win (#98)
