+++ momepy/graph.py
@@ -934,6 +934,9 @@ def local_betweenness_centrality(
     weight : str, optional
         Use the specified edge attribute as the edge distance in shortest
         path calculations in closeness centrality algorithm
+    normalized : bool, optional
+        If True the betweenness values are normalized by `2/((n-1)(n-2))`,
+        where n is the number of nodes in subgraph.
     **kwargs
         kwargs for ``networkx.betweenness_centrality_subset``
 
