### 2019-10-23 42f1ffd
BUG: generating angular graph with non-default index (#100)
* Update utils.py

* tst
