+++ tests/test_utils.py
@@ -42,6 +42,11 @@ class TestUtils:
         dual = mm.gdf_to_nx(self.df_streets, approach="dual")
         assert dual.number_of_nodes() == 35
         assert dual.number_of_edges() == 148
+        self.df_streets["ix"] = np.arange(0, len(self.df_streets) * 2, 2)
+        self.df_streets.set_index("ix", inplace=True)
+        dual2 = mm.gdf_to_nx(self.df_streets, approach="dual")
+        assert dual2.number_of_nodes() == 35
+        assert dual2.number_of_edges() == 148
         with pytest.raises(ValueError):
             mm.gdf_to_nx(self.df_streets, approach="nonexistent")
 
