### 2019-10-29 #1807 @drvinceknight
No news but both @geraintpalmer and @martibosch had indicated it might take them a while. @geraintpalmer @martibosch, very appreciative of your time: just pinging in case this slipped of your horizon.
