### 2019-10-29 #1807 @martibosch
Hello! sorry for my delay. I threw an eye on the library yesterday, but I have to get some other things done before friday. I will be getting to it momepy in more detail weekend, I hope this does not delay the reviewing process too much.
