### 2019-10-29 #1807 @martibosch
I am sorry about that. I have updated the comment, and I hope that the language used now is more appropriate. Thank you
