### 2019-10-29 #1807 @danielskatz
Thanks - I was just a bit uncertain how to read the original version - thanks again for working on this review.
