### 2019-10-29 8951667
BUG: _generate_dual() update for MultiGraph
