### 2019-10-29 8638dda
BUG: _generate_dual() update for MultiGraph (#101)
* BUG: _generate_dual() update for MultiGraph

* Fix tests

* fix mismatch
