### 2019-10-29 089a2fa
CI: use Python 3.8 in dev
