+++ .travis.yml
@@ -6,7 +6,7 @@ matrix:
   include:
     - env: ENV_FILE="ci/travis/37-pysal.yaml"
     - env: ENV_FILE="ci/travis/37-mc.yaml"
-    - env: ENV_FILE="ci/travis/37-dev.yaml" DEV=true
+    - env: ENV_FILE="ci/travis/38-dev.yaml" DEV=true
 
 install:
   - sudo apt-get update
