+++ ci/travis/38-dev.yaml
@@ -2,7 +2,7 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python=3.7
+  - python=3.8
   - geopandas
   - libpysal
   - networkx
