### 2019-10-29 be76468
BUG: PerimeterWall for non-consecutive index (#102)
