### 2019-11-01 martinfleis/momepy#103 @geraintpalmer
User Guide issues
Hi @martinfleis
I'm currently reviewing your JOSS paper. I've worked through the 'Gerring  Started' section of the user guide here: https://guide.momepy.org/getting_started.html

Some additional packages needed to be installed in order to complete the guide:

- `matplotlib` and `opsmx` (these was obvious from the guild, not a problem)
- `mapclassify` or `pysal` (these are listed as optional requirements, but are required for the 'getting started' section, and they are not installed when following the installation instructions (http://docs.momepy.org/en/latest/install.html). It's also not obvious these are optional requirements from the README or the requirements.txt file.
- `descartes` This was not mentioned anywhere.

I started a new conda environment (as instructed in the installation guide). Except for the above libraries, the only other package I installed was jupyter. I could not reproduce the following from the 'getting started' guide:

```
>>> profile = momepy.StreetProfile(streets, buildings)
```

Yielding the following error:

```
TypeError: unsupported operand type(s) for +: 'NoneType' and 'NoneType'
```
