### 2019-11-01 #1807 @geraintpalmer
Submitted issue https://github.com/martinfleis/momepy/issues/103 to the original repository.
