### 2019-11-01 martinfleis/momepy#103 @martinfleis
Hi @geraintpalmer,

thank you for working on this review. I'll expand dependencies instructions. `mapclassify` or `pysal` and `descartes` are required by `geopandas` not momepy (at least for this guide), but I'll comment on it.

`StreetProfile` issue has been fixed in #97 so it should not yield the error on master. 0.1rc4 release is older than this fix. You can try installing master if you want to (https://docs.momepy.org/en/latest/install.html#install-from-the-repository).
