### 2019-11-01 martinfleis/momepy#103 @martinfleis
Getting started expanded in https://github.com/martinfleis/momepy/commit/3ef3f5488639c254b792e3882fc9cacd1a6c4608.
It is already live https://guide.momepy.org/getting_started.html#dependencies

That should resolve all issues mentioned above (`TypeError` was resolved earlier). Closing as fixed. Feel free to open other issues if you'll find anything.

Thank you!
