### 2019-11-04 martinfleis/momepy#103 @geraintpalmer
Great thank you!

Yes if installing from master `StreetProfile` now works.
