### 2019-11-05 7f349e8
BUG: fix fractal_dimension formula
