+++ tests/test_shape.py
@@ -36,8 +36,10 @@ class TestShape:
 
     def test_FractalDimension(self):
         self.df_buildings["fd"] = mm.FractalDimension(self.df_buildings).fd
-        check = 2 * math.log(self.df_buildings.geometry[0].length / 4) / math.log(
-            self.df_buildings.geometry[0].area
+        check = (
+            2
+            * math.log(self.df_buildings.geometry[0].length / 4)
+            / math.log(self.df_buildings.geometry[0].area)
         )
         assert self.df_buildings["fd"][0] == check
 
