### 2019-11-05 273f3ba
BUG: fix FractalDimension formula (#104)
* BUG: fix fractal_dimension formula

* blacken
