### 2019-11-05 martinfleis/momepy#103 @martinfleis
Just pinging @martibosch here, as he might encounter the same issue.

The other way, without installing master, is to pass dummy building heights to `StreetProfile`. Or skipping it altogether :).

```py
import numpy as np
buildings["height"] = np.linspace(10.0, 30.0, 144)
profile = momepy.StreetProfile(streets, buildings, heights='height')
```
