### 2019-11-05 martinfleis/momepy#103 @geraintpalmer
I am having further issues following the contributon guidelines: https://docs.momepy.org/en/latest/contributing.html

When running `pytest` I get the following error (one of 8):

```
_______________________________________________ ERROR collecting tests/test_utils.py _______________________________________________
ImportError while importing test module '/Users/geraintpalmer/Documents/Personal/Reviews/momepy/tests/test_utils.py'.
Traceback:
tests/test_utils.py:1: in <module>
    import geopandas as gpd
//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/geopandas/__init__.py:5: in <module>
    from geopandas.io.file import read_file  # noqa
//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/geopandas/io/file.py:6: in <module>
    import fiona
//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/__init__.py:83: in <module>
    from fiona.collection import BytesCollection, Collection
//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/collection.py:9: in <module>
    from fiona.ogrext import Iterator, ItemsIterator, KeysIterator
E   ImportError: dlopen(//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/ogrext.cpython-37m-darwin.so, 2): Library not
loaded: @rpath/libpoppler.76.dylib
E     Referenced from: //anaconda3/envs/momepy_dev/lib/libgdal.20.dylib
E     Reason: image not found
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 8 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
========================================================= 8 error in 0.67s =========================================================

```

I belive this is to do with the requirement `fiona`, as when I import fiona into python I get:

```
>>> import fiona
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/__init__.py", line 83, in <module>
    from fiona.collection import BytesCollection, Collection
  File "//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/collection.py", line 9, in <module>
    from fiona.ogrext import Iterator, ItemsIterator, KeysIterator
ImportError: dlopen(//anaconda3/envs/momepy_dev/lib/python3.7/site-packages/fiona/ogrext.cpython-37m-darwin.so, 2): Library not load
ed: @rpath/libpoppler.76.dylib
  Referenced from: //anaconda3/envs/momepy_dev/lib/libgdal.20.dylib
  Reason: image not found
```

Could you advise please?
