### 2019-11-05 martinfleis/momepy#103 @martinfleis
That looks like a Fiona issue caused by mixed-source environment. How did you install geopandas/fiona into this env? I assume it is the different one than you have used to run Getting Started.
