### 2019-11-05 martinfleis/momepy#106 @martibosch
AttributeError in links.ipynb (from user guide)
I wonder whether this is related to my version of networkx (2.4) or any other library (feel free to ask for the whole list of packages of my conda environment), but I am unable to reproduce the `links.ipynb` notebook from the user guide. More precisely, the seventh code-cell raises the following error:

```python>
nodes, edges = momepy.nx_to_gdf(graph)

Truncated Traceback (Use C-c C-$ to view full TB):
~/anaconda3/envs/momepy/lib/python3.7/site-packages/momepy/utils.py in _lines_to_gdf(net, lines, points, nodeID)
    240         node_end = []
    241         for s in starts:
--> 242             node_start.append(net.node[s][nodeID])
    243         for e in ends:
    244             node_end.append(net.node[e][nodeID])

AttributeError: 'MultiGraph' object has no attribute 'node'
```
I wonder whether this is just a typo, since networkx `Graph` and `MultiGraph` objects do have a `nodes` property.
