### 2019-11-05 #1807 @martibosch
Submitted issue [martinfleis/momepy#106](https://github.com/martinfleis/momepy/issues/106) to the original repository.
