### 2019-11-05 martinfleis/momepy#103 @martinfleis
I think I know where is the issue. Following the instructions, you get everything from conda defaults, instead of conda-forge. That sometimes cause fiona errors. I’ll amend instructions to use conda-forge.
