### 2019-11-05 martinfleis/momepy#106 @martinfleis
Yes, it is related to networkx 2.4 as they deprecated it. I already fixed that in #96, so master should be okay.
