### 2019-11-05 martinfleis/momepy#103 @geraintpalmer
Yes this was a new environment following the contributiing instructions. I did pip install pytest rather than conda install it though.
