### 2019-11-05 9e89489
Typos
Fixes #105
