### 2019-11-05 martinfleis/momepy#103 @martinfleis
pytest is okay, the problem is usually with fiona's GDAL dependency which is C library. However, I am struggling to replicate this error as following the instructions I get all packages from conda-forge. Can you show `conda list` for this environment?

This is might be the better way how to set-up development env, but on my Mac and linux machines the resulting environments are the same:

```
conda create -n momepy_dev
conda activate momepy_dev
conda config --env --add channels conda-forge
conda config --env --set channel_priority strict
conda install geopandas networkx libpysal tqdm pysal mapclassify
```
