### 2019-11-05 #1807 @martibosch
A comment (part of my review), here (I believe raising a dedicated issue is not so appropriate).

I wonder whether (I say "wonder" because I am sure that the author @martinfleis has given it much more thought than I have) the object-oriented approach to the computation of characters (especially the `dimension` module, but also the `shape`, `intensity` and `diversity` modules).
Most of the classes consist of only the `__init__` method, whose main purpose is to return a pandas series. I believe that such design is less intuitive for the end user. For instance, if I check the output of `tess_car` in a Python interpreter, I get `<momepy.intensity.AreaRatio at 0x7f37a235a240>`. I am most likely interested in the pandas series of area ratios of my tesselation, but why should I know that this is set as its `ar` attribute? Furthermore, since there are many characters implemented in momepy, the users need to know the respective attribute name each time (it seems that  I'd suggest: either implementing the `__repr__` method so that the user gets more information, or make them accessible through an attribute that has the same name accross each characters' class (e.g., `series`).

* the `__init__` method also sets some other attributes, but they are often just taken directly from the provided arguments, or result of very straight-forward computations (e.g., [this line](https://github.com/martinfleis/momepy/blob/master/momepy/dimension.py#L149).
* some classes have methods other than `__init__` (e.g., [this](https://github.com/martinfleis/momepy/blob/master/momepy/dimension.py#L323-L325) or [this](https://github.com/martinfleis/momepy/blob/6a31c1660d32dd7da61fbc7634392f2cc0e4e443/momepy/diversity.py#L306-L324)), but I have not seen any that could not be implemented outside the class.

Therefore, I see no gains from the object-oriented approach (again, let me re-state that this is just my first impression and that I might be missing many possibilities that justify the author's adoption of the object-oriented design), other than design consistency, e.g., with the more convoluted `elements.py` module.
Of course, consistency is desirable, but I wonder whether the object-oriented design is the best approach for this case. From what I understand, users are interested in the series of computed characters, thus it might be more appropriate to implement the computation of characters as functions instead, e.g.:

```python
def floor_area(gdf, heights, areas=None):
    # code to compute `fa` here
    return fa
```

Of course I do not expect the author to change the code organisation at this point (and again, there might be good reasons for the object-oriented approach that I am missing), but I just wanted to raise the question.
