### 2019-11-05 #1807 @martinfleis
@martibosch Thank you for this comment. It is very valid and there has been a lot of discussion on this topic as it was originally designed as you propose - it simply returned pandas Series. However, this has been changed to object-oriented for several reasons.

First is the consistency within the package as some classes returns more than one series (like `StreetProfile`). Good point is the possible implementation of `__repr__` and consistent attribute to get the resulting series (if there is only one), I'll certainly look to do both in the future. Thank you for that idea.

Second is the consistency with the PySAL, which is very close to the scope of momepy. When discussing how should be momepy designed, I mostly adopted their object-oriented approach (example - https://github.com/pysal/inequality/blob/master/inequality/gini.py).

As I expect momepy to grow, I also wanted to keep our classes more flexible and extensible than functions are. I expect that some more complex classes might get another attributes and it is preferred to keep them within one class, rather than having individual functions. Some surely won't, but then we, again, want consistency.

Finally, classes also store parameters used for each computation. Some of them generate additional information, for example, spatial weights or perimeter values under the hood. I see the value in the possibility to access them (as they can be reused - spatial weight might be expensive to compute - or explored). Returning them alongside the resulting series just complicates the design.

I have to admit that this decision was not straightforward, but in the end I felt that object-oriented approach is a bit more favourable here. But I am open to discussion if the decision was right.
