### 2019-11-05 #1807 @martibosch
@martinfleis thank you for the explanations, which have been very convincing. Again, you indeed know momepy way better than I do, so after your explanations, I fully trust your criteria. As a reviewer, I just wanted to make sure that such matters had been properly consider - which I know by now, so as far as I am concerner, I believe we can move on with the review.
I have a couple more issues that I will submit tomorrow and then I will proceed to reviewing the manuscript. Cheers!
