### 2019-11-05 7bb4bf3
API: unify results series attribute
