+++ tests/test_shape.py
@@ -15,10 +15,10 @@ class TestShape:
         self.df_streets = gpd.read_file(test_file_path, layer="streets")
         self.df_tessellation = gpd.read_file(test_file_path, layer="tessellation")
         self.df_buildings["height"] = np.linspace(10.0, 30.0, 144)
-        self.df_buildings["volume"] = mm.Volume(self.df_buildings, "height").volume
+        self.df_buildings["volume"] = mm.Volume(self.df_buildings, "height").series
 
     def test_FormFactor(self):
-        self.df_buildings["ff"] = mm.FormFactor(self.df_buildings, "volume").ff
+        self.df_buildings["ff"] = mm.FormFactor(self.df_buildings, "volume").series
         check = (self.df_buildings.geometry[0].area) / (
             self.df_buildings.volume[0] ** (2 / 3)
         )
@@ -26,16 +26,16 @@ class TestShape:
 
         self.df_buildings["ff"] = mm.FormFactor(
             self.df_buildings,
-            mm.Volume(self.df_buildings, "height").volume,
+            mm.Volume(self.df_buildings, "height").series,
             areas=self.df_buildings.geometry.area,
-        ).ff
+        ).series
         check = (self.df_buildings.geometry[0].area) / (
             self.df_buildings.volume[0] ** (2 / 3)
         )
         assert self.df_buildings["ff"][0] == check
 
     def test_FractalDimension(self):
-        self.df_buildings["fd"] = mm.FractalDimension(self.df_buildings).fd
+        self.df_buildings["fd"] = mm.FractalDimension(self.df_buildings).series
         check = (
             2
             * math.log(self.df_buildings.geometry[0].length / 4)
@@ -47,30 +47,30 @@ class TestShape:
             self.df_buildings,
             areas=self.df_buildings.geometry.area,
             perimeters=self.df_buildings.geometry.length,
-        ).fd
+        ).series
         assert self.df_buildings["fd2"][0] == check
 
     def test_VolumeFacadeRatio(self):
         self.df_buildings["peri"] = self.df_buildings.geometry.length
         self.df_buildings["vfr"] = mm.VolumeFacadeRatio(
             self.df_buildings, "height", "volume", "peri"
-        ).vfr
+        ).series
         check = self.df_buildings.volume[0] / (
             self.df_buildings.peri[0] * self.df_buildings.height[0]
         )
         assert self.df_buildings["vfr"][0] == check
 
         peri = self.df_buildings.geometry.length
-        volume = mm.Volume(self.df_buildings, "height").volume
+        volume = mm.Volume(self.df_buildings, "height").series
         self.df_buildings["vfr2"] = mm.VolumeFacadeRatio(
             self.df_buildings, "height", volume, peri
-        ).vfr
+        ).series
         assert self.df_buildings["vfr2"][0] == check
 
         self.df_buildings["peri"] = self.df_buildings.geometry.length
         self.df_buildings["vfr3"] = mm.VolumeFacadeRatio(
             self.df_buildings, "height"
-        ).vfr
+        ).series
 
         assert self.df_buildings["vfr3"][0] == check
 
@@ -78,7 +78,7 @@ class TestShape:
         self.df_buildings["area"] = self.df_buildings.geometry.area
         self.df_buildings["circom"] = mm.CircularCompactness(
             self.df_buildings, "area"
-        ).cc
+        ).series
         check = self.df_buildings.area[0] / (
             _circle_area(
                 list(self.df_buildings.geometry[0].convex_hull.exterior.coords)
@@ -89,14 +89,14 @@ class TestShape:
         area = self.df_buildings.geometry.area
[…]
