+++ tests/test_dimension.py
@@ -18,12 +18,12 @@ class TestDimensions:
         self.df_buildings["height"] = np.linspace(10.0, 30.0, 144)
 
     def test_Area(self):
-        self.df_buildings["area"] = mm.Area(self.df_buildings).area
+        self.df_buildings["area"] = mm.Area(self.df_buildings).series
         check = self.df_buildings.geometry[0].area
         assert self.df_buildings["area"][0] == check
 
     def test_Perimeter(self):
-        self.df_buildings["perimeter"] = mm.Perimeter(self.df_buildings).perimeter
+        self.df_buildings["perimeter"] = mm.Perimeter(self.df_buildings).series
         check = self.df_buildings.geometry[0].length
         assert self.df_buildings["perimeter"][0] == check
 
@@ -31,17 +31,17 @@ class TestDimensions:
         self.df_buildings["area"] = self.df_buildings.geometry.area
         self.df_buildings["volume"] = mm.Volume(
             self.df_buildings, "height", "area"
-        ).volume
+        ).series
         check = self.df_buildings.geometry[0].area * self.df_buildings.height[0]
         assert self.df_buildings["volume"][0] == check
 
         area = self.df_buildings.geometry.area
         height = np.linspace(10.0, 30.0, 144)
-        self.df_buildings["volume"] = mm.Volume(self.df_buildings, height, area).volume
+        self.df_buildings["volume"] = mm.Volume(self.df_buildings, height, area).series
         check = self.df_buildings.geometry[0].area * self.df_buildings.height[0]
         assert self.df_buildings["volume"][0] == check
 
-        self.df_buildings["volume"] = mm.Volume(self.df_buildings, "height").volume
+        self.df_buildings["volume"] = mm.Volume(self.df_buildings, "height").series
         check = self.df_buildings.geometry[0].area * self.df_buildings.height[0]
         assert self.df_buildings["volume"][0] == check
 
@@ -54,7 +54,7 @@ class TestDimensions:
         self.df_buildings["area"] = self.df_buildings.geometry.area
         self.df_buildings["floor_area"] = mm.FloorArea(
             self.df_buildings, "height", "area"
-        ).fa
+        ).series
         check = self.df_buildings.geometry[0].area * (self.df_buildings.height[0] // 3)
         assert self.df_buildings["floor_area"][0] == check
 
@@ -62,10 +62,12 @@ class TestDimensions:
         height = np.linspace(10.0, 30.0, 144)
         self.df_buildings["floor_area"] = mm.FloorArea(
             self.df_buildings, height, area
-        ).fa
+        ).series
         assert self.df_buildings["floor_area"][0] == check
 
-        self.df_buildings["floor_area"] = mm.FloorArea(self.df_buildings, "height").fa
+        self.df_buildings["floor_area"] = mm.FloorArea(
+            self.df_buildings, "height"
+        ).series
         assert self.df_buildings["floor_area"][0] == check
 
         with pytest.raises(KeyError):
@@ -77,7 +79,7 @@ class TestDimensions:
         self.df_buildings["area"] = self.df_buildings.geometry.area
         self.df_buildings["courtyard_area"] = mm.CourtyardArea(
             self.df_buildings, "area"
-        ).ca
+        ).series
         check = (
             Polygon(self.df_buildings.geometry[80].exterior).area
             - self.df_buildings.geometry[80].area
@@ -87,10 +89,10 @@ class TestDimensions:
         area = self.df_buildings.geometry.area
         self.df_buildings["courtyard_area"] = mm.CourtyardArea(
             self.df_buildings, area
-        ).ca
+        ).series
[…]

[…]
