### 2019-11-06 martinfleis/momepy#108 @martibosch
import of `_longest_axis`
Following [the review comment](https://github.com/openjournals/joss-reviews/issues/1807#issuecomment-549947746) that I left of the main JOSS thread, I noticed that [the import of `_longest_axis` in `shape.py`](
https://github.com/martinfleis/momepy/blob/9e894893bf731126aa8cbd009b459374a390b43b/momepy/shape.py#L1177) should rasie an `ImportError` in any case, since such function exists only as the `_longest_axis` method of `LongestAxisLength`.

It seems that the import would only be effective when all of the buildings of `gdf` are of circular geometry. While this is very unlikely to happen, in the case that it does, it will certainly raise an `ImportError`.
