### 2019-11-06 martinfleis/momepy#109 @martibosch
`SettingWithCopyWarning` in the intensity notebook/user guide
In the `intensity.ipynb` notebook , the call to:
```python
tess_car = momepy.AreaRatio(tessellation, buildings, 'area', 'area', 'uID')
```
throws a pandas `SettingWithCopyWarning` as in:

```
/home/martibosch/anaconda3/envs/momepy/lib/python3.7/site-packages/pandas/core/frame.py:4238: SettingWithCopyWarning:
A value is trying to be set on a copy of a slice from a DataFrame

See the caveats in the documentation: http://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
  return super().rename(**kwargs)
```
and the same happens for:

```python
tessellation['FAR'] = momepy.AreaRatio(tessellation, buildings,
                                                     'area', 'floor_area', 'uID').ar
```

```
/home/martibosch/anaconda3/envs/momepy/lib/python3.7/site-packages/pandas/core/frame.py:4238: SettingWithCopyWarning:
A value is trying to be set on a copy of a slice from a DataFrame

See the caveats in the documentation: http://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
  return super().rename(**kwargs)
```

Given how the classes to compute characters work, it is likely that the `SettingWithCopyWarning` occurs elsewhere. While this is just a warning, I believe that it is generally a good practice to avoid it (if possible).
