### 2019-11-06 #1807 @martibosch
Submitted issue [martinfleis/momepy#109](https://github.com/martinfleis/momepy/issues/109) to the original repository
