### 2019-11-06 martinfleis/momepy#110 @martibosch
`TypeError` in the `distribution.ipynb` notebook/user guide
The following line of the `distribution.ipynb` notebook:
```python
buildings['networkID'] = momepy.get_network_id(buildings, edges,
                                               'uID', 'networkID')
```
raises a `TypeError` with the following full traceback:

```
<ipython-input-20-2d62b5588232> in <module>
      1 buildings['networkID'] = momepy.get_network_id(buildings, edges,
----> 2                                                'uID', 'networkID')

~/anaconda3/envs/momepy/lib/python3.7/site-packages/momepy/elements.py in get_network_id(left, right, network_id, min_size)
    691     ):
    692         p = r.geometry
--> 693         pbox = (p.x - MIN_SIZE, p.y - MIN_SIZE, p.x + MIN_SIZE, p.y + MIN_SIZE)
    694         hits = list(idx.intersection(pbox))
    695         d = INFTY

TypeError: unsupported operand type(s) for -: 'float' and 'str'
```

The positional arguments of the `get_network_id` method are clearly mismatched. It should be easy to fix. It is perhaps a good idea to add a test that prevents that.
