### 2019-11-06 #1807 @martibosch
Submitted issue [martinfleis/momepy#110](https://github.com/martinfleis/momepy/issues/110) to the original repository
