### 2019-11-06 martinfleis/momepy#111 @martibosch
paper corrections
I have two quick notes on the paper. I should note that I am not a native English speaker so I might be wrong in some parts.

* I believe that the sentence that starts by "Under the hood uses PySAL..." is semantically incorrect: if I am not mistaken, in the way the sentence is written, the only possible subject is the "uses" word (if understood as substantive). However, the subject should be "momepy", e.g., "momepy uses PySAL under the hood, mostly to take care of spatial weights ..."

* The "exemplar Jupyter notebooks" may be technically correct, but I believe that for many readers, it carries the connotations of "ideal". I would use "example Jupyter notebooks" instead (this might be personal taste)
