### 2019-11-06 #1807 @martibosch
Submitted issue [martinfleis/momepy#111](https://github.com/martinfleis/momepy/issues/111) to the original repository
