### 2019-11-06 martinfleis/momepy#103 @geraintpalmer
Here's the output of `conda list`. Note I pip install pytest after conda installing everyting else.

```
# packages in environment at //anaconda3/envs/momepy_dev2:
#
# Name                    Version                   Build  Channel
atomicwrites              1.3.0                    pypi_0    pypi
attrs                     19.3.0                     py_0    conda-forge
boost-cpp                 1.70.0               hd59e818_1    conda-forge
bzip2                     1.0.8                h01d97ff_1    conda-forge
ca-certificates           2019.9.11            hecc5488_0    conda-forge
cairo                     1.16.0            h0ab9d94_1001    conda-forge
certifi                   2019.9.11                py37_0    conda-forge
click                     7.0                      pypi_0    pypi
click-plugins             1.1.1                      py_0    conda-forge
cligj                     0.5.0                    pypi_0    pypi
curl                      7.65.3               ha441bb4_0
cycler                    0.10.0                     py_2    conda-forge
decorator                 4.4.1                      py_0    conda-forge
deprecated                1.2.6                      py_0    conda-forge
descartes                 1.1.0                      py_4    conda-forge
expat                     2.2.6                h0a44026_0
fiona                     1.8.4            py37h9a122fd_0
fontconfig                2.13.1            h1027ab8_1000    conda-forge
freetype                  2.10.0               h24853df_1    conda-forge
freexl                    1.0.5             h1de35cc_1002    conda-forge
gdal                      2.3.3            py37hbe65578_0
geopandas                 0.6.1                      py_1    conda-forge
geos                      3.7.1             h0a44026_1000    conda-forge
gettext                   0.19.8.1          h46ab8bc_1002    conda-forge
giflib                    5.1.7                h01d97ff_1    conda-forge
glib                      2.58.3            h9d45998_1002    conda-forge
hdf4                      4.2.13            hf3c6af0_1002    conda-forge
hdf5                      1.10.4          nompi_h0cbb7df_1106    conda-forge
icu                       58.2              h0a44026_1000    conda-forge
importlib-metadata        0.23                     pypi_0    pypi
joblib                    0.14.0                     py_0    conda-forge
jpeg                      9c                h1de35cc_1001    conda-forge
json-c                    0.13.1            h1de35cc_1001    conda-forge
kealib                    1.4.10            hecf890f_1003    conda-forge
kiwisolver                1.1.0            py37h770b8ee_0    conda-forge
krb5                      1.16.1               hddcf347_7
libblas                   3.8.0               14_openblas    conda-forge
libcblas                  3.8.0               14_openblas    conda-forge
libcurl                   7.65.3               h051b688_0
libcxx                    4.0.1                hcfea43d_1
libcxxabi                 4.0.1                hcfea43d_1
libdap4                   3.19.1               h3d3e54a_0
libedit                   3.1.20181209         hb402a30_0
libffi                    3.2.1                h475c297_4
libgdal                   2.3.3                h0950a36_0
libgfortran               3.0.1                         0    conda-forge
libiconv                  1.15              h01d97ff_1005    conda-forge
libkml                    1.3.0             hed7d534_1010    conda-forge
liblapack                 3.8.0               14_openblas    conda-forge
libnetcdf                 4.6.2             h6b88ef6_1001    conda-forge
libopenblas               0.3.7                hd44dcd8_1    conda-forge
libpng                    1.6.37               h2573ce8_0    conda-forge
libpq                     11.2                 h051b688_0
libpysal                  4.1.1                      py_0    conda-forge
libspatialindex           1.8.5                h2c08c6b_2
libspatialite             4.3.0a            h0cd9627_1026    conda-forge
libssh2                   1.8.2                hcdc9a53_2    conda-forge
libtiff                   4.0.10            hd08fb8f_1003    conda-forge
libxml2                   2.9.9                hf6e021a_1
llvm-openmp               9.0.0                h40edb58_0    conda-forge
lz4-c                     1.8.3             h6de7cb9_1001    conda-forge
mapclassify               2.1.1                      py_0    conda-forge
matplotlib                3.1.1                    py37_1    conda-forge
matplotlib-base           3.1.1            py37h3a684a6_1    conda-forge
momepy                    0.1rc4                    dev_0    <develop>
more-itertools            7.2.0                    pypi_0    pypi
munch                     2.5.0                      py_0    conda-forge
ncurses                   6.1                  h0a44026_1
networkx                  2.4                        py_0    conda-forge
numpy                     1.17.3           py37hde6bac1_0    conda-forge
openjpeg                  2.3.1                hdc36067_1    conda-forge
openssl                   1.1.1c               h01d97ff_0    conda-forge
packaging                 19.2                     pypi_0    pypi
palettable                3.3.0                      py_0    conda-forge
pandas                    0.24.2           py37h86efe34_0    conda-forge
patsy                     0.5.1                      py_0    conda-forge
pcre                      8.43                 h0a44026_0
pip                       19.3.1                   py37_0    conda-forge
pixman                    0.38.0            h01d97ff_1003    conda-forge
pluggy                    0.13.0                   pypi_0    pypi
poppler                   0.81.0               h66b6897_0    conda-forge
poppler-data              0.4.9                         1    conda-forge
proj4                     5.2.0             h6de7cb9_1006    conda-forge
py                        1.8.0                    pypi_0    pypi
pyparsing                 2.4.4                      py_0    conda-forge
pyproj                    1.9.6            py37h9c430a6_0
pysal                     2.1.0                      py_0    conda-forge
pytest                    5.2.2                    pypi_0    pypi
python                    3.7.4                h359304d_1
python-dateutil           2.8.1                      py_0    conda-forge
pytz                      2019.3                     py_0    conda-forge
readline                  7.0                  h1de35cc_5
rtree                     0.8.3                    py37_0
scikit-learn              0.21.3           py37hd4ffd6c_0    conda-forge
scipy                     1.3.1            py37hab3da7d_2    conda-forge
seaborn                   0.9.0                      py_1    conda-forge
setuptools                41.6.0                   py37_1    conda-forge
shapely                   1.6.4           py37h79c6f3e_1005    conda-forge
six                       1.13.0                   py37_0    conda-forge
sqlite                    3.30.1               ha441bb4_0
statsmodels               0.10.1           py37h3b54f70_1    conda-forge
tk                        8.6.8                ha441bb4_0
tornado                   6.0.3            py37h0b31af3_0    conda-forge
tqdm                      4.36.1                     py_0    conda-forge
wcwidth                   0.1.7                    pypi_0    pypi
wheel                     0.33.6                   py37_0    conda-forge
wrapt                     1.11.2           py37h0b31af3_0    conda-forge
xerces-c                  3.2.2             h4db8090_1003    conda-forge
xz                        5.2.4                h1de35cc_4
zipp                      0.6.0                    pypi_0    pypi
zlib                      1.2.11               h1de35cc_3
zstd                      1.4.0                ha9f0a20_0    conda-forge
```
