### 2019-11-06 #1807 @martibosch
This should be all from my side. I would just like to add a final note which I do not intend to be part of the reviewing. The documentation is provided as a Jupyter book that is hosted at a separate `gh-pages` branch of the repository. This is the first time that I see such an approach, and I have some concerns about its maintainability, since unless I am missing something, it seems that two independent git branches (that will never be merged) are hosted within the same repo. Wouldn't a separate repository, such as [osmnx-examples](https://github.com/gboeing/osmnx-examples) be a better approach? In any case, I believe that a [binder](https://mybinder.org/) badge would be a nice addition.
On the other hand, the notebooks of the user guide could be added as part of the Travis CI workflow (even with the current single repository with a separate `gh-pages` branch), so that the example notebooks run properly in each commit. This could potentially avoid some of the issues that I have encountered when running them and add extra robustness to the library. At the risk of "shameless self promotion", [the travis setup of PyLandStats](https://github.com/martibosch/pylandstats/blob/master/.travis.yml) and [this Makefile](https://github.com/martibosch/pylandstats-notebooks/blob/master/Makefile) might be of help.

In any case, I believe that this is a great package that paves the way for automated and reproducible analyses of urban morphology at an unprecedented scale.
