### 2019-11-06 martinfleis/momepy#103 @martinfleis
@geraintpalmer Thank you. The issue here is that fiona is installed from conda defaults for some reason. And it is not even the latest one. If you try the other approach (previous comment), does the issue persist? I would change the contributing guidelines to follow that one if that works.
