### 2019-11-06 #1807 @martinfleis
> Wouldn't a separate repository, such as osmnx-examples be a better approach? In any case, I believe that a binder badge would be a nice addition.

I have to admit that such an obvious idea did not came to my mind. Absolutely, I'll do that once I'll find some spare time. I have tried to make binder work but gave up after a while. It is on my to-do list since then.

> the notebooks of the user guide could be added as part of the Travis CI workflow

That is a good point. I'll check your repo for an inspiration and include them (once migrated to separate repository).

Thank you for you thorough review. It is extremely helpful. I'll try to deal with the issues you have opened ASAP.
