### 2019-11-06 b18e4a9
BUG: import of `_longest_axis`
