### 2019-11-06 martinfleis/momepy#108 @martinfleis
Thanks, that is indeed a bug.

It would import `_longest_axis` when at least one building is circular as is iterates through geometries. Fixed in #112
