### 2019-11-06 martinfleis/momepy#109 @martinfleis
Thank you! Fixed in #113

> Given how the classes to compute characters work, it is likely that the SettingWithCopyWarning occurs elsewhere.

It does not, pytest raised this Warning only for `AreaRatio`.
