### 2019-11-06 martinfleis/momepy#110 @martinfleis
Oh, yes. `'uID'` should not be there. Thanks! (I certainly need to put User Guide to Travis.)
