### 2019-11-06 martinfleis/momepy#111 @martinfleis
Thank you. I reshuffled affected sentences in https://github.com/martinfleis/momepy/commit/b868e4bf09a10299a646a0fa4e4e7a547348fc48.
