### 2019-11-06 96f0d03
BUG: import of `_longest_axis` (#112)
* BUG: import of `_longest_axis`

* remove reimport
