### 2019-11-06 #1807 @martinfleis
@martibosch All issues you have opened should be fixed and merged into the master now. Thanks again!
