### 2019-11-06 martinfleis/momepy#103 @martinfleis
After resetting my conda settings to default I can reproduce this behaviour.

Now I can confirm that instructions I have posted above works, unlike those in contributing docs (if you did not set channel_priority to strict globally). I'll change that in documentation. Thanks!
