### 2019-11-07 martinfleis/momepy#103 @geraintpalmer
Fantastic thank you @martinfleis that all worked!
