### 2019-11-09 #1807 @martibosch
@martinfleis sorry that I forgot about the checkboxes. There is only one final checkbox that I am still leaving unchecked: not all the references have a DOI, which conflicts with [the review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper).
