### 2019-11-09 #1807 @martinfleis
@martibosch those without DOI do not have one at all.
