### 2019-11-09 #1807 @martibosch
in such case, I believe that I can put a check in the “References” checkbox, am I wrong @drvinceknight?
