### 2019-11-10 03bb1d5
update travis to fail on black
