### 2019-11-10 373fa4b
CI config (#115)
* AppVeyor config

* badge

* update travis to fail on black

* style
