+++ momepy/dimension.py
@@ -321,8 +321,7 @@ class LongestAxisLength:
     def __init__(self, gdf):
         self.gdf = gdf
         self.lal = gdf.apply(
-            lambda row: _longest_axis(row.geometry.convex_hull.exterior.coords),
-            axis=1,
+            lambda row: _longest_axis(row.geometry.convex_hull.exterior.coords), axis=1
         )
 
 
