+++ setup.cfg
@@ -1,0 +1,11 @@
+# See the docstring in versioneer.py for instructions. Note that you must
+# re-run 'versioneer.py setup' after changing this section, and commit the
+# resulting files.
+
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = momepy/_version.py
+versionfile_build = momepy/_version.py
+tag_prefix = v
+parentdir_prefix = momepy-
