+++ momepy/__init__.py
@@ -12,5 +12,6 @@ __author__ = "Martin Fleischmann"
 __author_email__ = "martin@martinfleischmann.net"
 
 from ._version import get_versions
-__version__ = get_versions()['version']
+
+__version__ = get_versions()["version"]
 del get_versions
