+++ setup.cfg
@@ -1,7 +1,3 @@
-# See the docstring in versioneer.py for instructions. Note that you must
-# re-run 'versioneer.py setup' after changing this section, and commit the
-# resulting files.
-
 [versioneer]
 VCS = git
 style = pep440
