### 2019-11-10 431c495
ENH: Versioneer (#116)
* use versioneer

* black

* clean

* codecov ignore
