### 2019-11-10 c09ed7b
make readthedocs lightweight
