+++ docs/environment.yml
@@ -4,17 +4,7 @@ channels:
 dependencies:
     - sphinx_bootstrap_theme=0.6.4
     - python=3.7
-    - geopandas
-    - libpysal
-    - networkx
-    - tqdm
-    - mapclassify
-    - momepy
-    - matplotlib
-    - rtree
-    - osmnx
     - pip
     - pip:
-        - inequality
         - sphinxcontrib.bibtex
         - numpydoc
