+++ momepy/elements.py
@@ -14,7 +14,12 @@ from scipy.spatial import Voronoi
 from shapely.geometry import MultiPolygon, Point, Polygon
 from shapely.wkt import loads
 from tqdm import tqdm
-from osgeo import ogr
+
+try:
+    from osgeo import ogr
+except ModuleNotFoundError:
+    # nothing - temporary solution for readthedocs fail.
+    print('')
 
 __all__ = ["buffered_limit", "Tessellation", "Blocks", "get_network_id", "get_node_id"]
 
