+++ momepy/elements.py
@@ -15,10 +15,10 @@ from shapely.geometry import MultiPolygon, Point, Polygon
 from shapely.wkt import loads
 from tqdm import tqdm
 
+# nothing - temporary solution for readthedocs fail. - cannot mock osgeo
 try:
     from osgeo import ogr
 except ModuleNotFoundError:
-    # nothing - temporary solution for readthedocs fail.
     print("rtd")
 
 __all__ = ["buffered_limit", "Tessellation", "Blocks", "get_network_id", "get_node_id"]
