+++ docs/conf.py
@@ -36,9 +36,6 @@ autodoc_mock_imports = [
     "mapclassify",
     "osmnx",
     "inequality",
-    "gdal",
-    "ogr"
-    "osgeo",
 ]
 
 # -- Project information -----------------------------------------------------
