### 2019-11-10 429f7b9
DOC: use mock for rtd (#117)
* make readthedocs lightweight

* import momepy

* mock

* mock osgeo?

* try osgeo

* matplotlib

* exclude from codecov

* comment

* clean
