+++ setup.cfg
@@ -5,3 +5,7 @@ versionfile_source = momepy/_version.py
 versionfile_build = momepy/_version.py
 tag_prefix = v
 parentdir_prefix = momepy-
+
+[flake8]
+# Black enforces 88 characters line length
+max_line_length = 88
