### 2019-11-10 4f0465a
DOC: change print in BuildingAdjacency
