### 2019-11-11 #1807 @geraintpalmer
I am happy with the paper now. Thanks @martinfleis for working through the issues.
