+++ momepy/dimension.py
@@ -320,8 +320,7 @@ class LongestAxisLength:
     def __init__(self, gdf):
         self.gdf = gdf
         self.series = gdf.apply(
-            lambda row: _longest_axis(row.geometry.convex_hull.exterior.coords),
-            axis=1,
+            lambda row: _longest_axis(row.geometry.convex_hull.exterior.coords), axis=1
         )
 
 
