### 2019-11-11 b1d8cba
API: unify results series attribute (#107)
* API: unify results series attribute

* fix

* blacken
