### 2019-11-11 9eb1f34
DOC: update contributing guidelines
