### 2019-11-11 a6eb193
DOC: update contributing guidelines (#118)
