### 2019-11-12 #1807 @drvinceknight
> in such case, I believe that I can put a check in the “References” checkbox, am I wrong @drvinceknight?

@martinfleis are there other alternative references with DOIs that you could include (not necessarily instead of but perhaps with)?
