### 2019-11-12 2c8fea0
scipy reference
