+++ paper/paper.bib
@@ -178,10 +178,30 @@ pages = {131--144},
 isbn = {90-8594-002-8},
 }
 
-@Misc{SciPy:suv-1Mya,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
-  url = {http://www.scipy.org/},
-  note = {[Online; accessed 2019-10-09]}
+@article{SciPy:suv-1Mya,
+       author = {{Virtanen}, Pauli and {Gommers}, Ralf and {Oliphant},
+         Travis E.  and {Haberland}, Matt and {Reddy}, Tyler and
+         {Cournapeau}, David and {Burovski}, Evgeni and {Peterson}, Pearu
+         and {Weckesser}, Warren and {Bright}, Jonathan and {van der Walt},
+         St{\'e}fan J.  and {Brett}, Matthew and {Wilson}, Joshua and
+         {Jarrod Millman}, K.  and {Mayorov}, Nikolay and {Nelson}, Andrew
+         R.~J. and {Jones}, Eric and {Kern}, Robert and {Larson}, Eric and
+         {Carey}, CJ and {Polat}, {\.I}lhan and {Feng}, Yu and {Moore},
+         Eric W. and {Vand erPlas}, Jake and {Laxalde}, Denis and
+         {Perktold}, Josef and {Cimrman}, Robert and {Henriksen}, Ian and
+         {Quintero}, E.~A. and {Harris}, Charles R and {Archibald}, Anne M.
+         and {Ribeiro}, Ant{\^o}nio H. and {Pedregosa}, Fabian and
+         {van Mulbregt}, Paul and {Contributors}, SciPy 1. 0},
+        title = "{SciPy 1.0--Fundamental Algorithms for Scientific
+                  Computing in Python}",
+      journal = {arXiv e-prints},
+         year = "2019",
+        month = "Jul",
+          eid = {arXiv:1907.10121},
+        pages = {arXiv:1907.10121},
+archivePrefix = {arXiv},
+       eprint = {1907.10121},
+ primaryClass = {cs.MS},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190710121V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
