### 2019-11-12 #1807 @martinfleis
@drvinceknight I am not aware of any. I looked for them when I wrote the paper, this is the best I was able to find. I have changed SciPy reference based on https://www.scipy.org/citing.html but that is probably all I can do.
