### 2019-11-13 64d444c
CI: Use python 3.8, allow failures in DEV
