+++ .travis.yml
@@ -2,7 +2,7 @@ language: python
 
 sudo: false
 
-matrix:
+jobs:
   include:
     - env: ENV_FILE="ci/travis/37-pysal.yaml"
     - env: ENV_FILE="ci/travis/37-mc.yaml" STYLE=true
