+++ .travis.yml
@@ -2,10 +2,11 @@ language: python
 
 sudo: false
 
-jobs:
+matrix:
   include:
     - env: ENV_FILE="ci/travis/37-pysal.yaml"
     - env: ENV_FILE="ci/travis/37-mc.yaml" STYLE=true
+    - env: ENV_FILE="ci/travis/dev.yaml" DEV=true
   allow_failures:
     - env: ENV_FILE="ci/travis/dev.yaml" DEV=true
 
