+++ ci/travis/dev.yaml
@@ -15,3 +15,4 @@ dependencies:
     - black
     - nose
     - mapclassify==2.0.1
+    - inequality
