+++ .github/workflows/momepy_CI.yml
@@ -1,0 +1,32 @@
+name: momepy CI
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      max-parallel: 9
+      matrix:
+        os: [macos-latest, windows-latest, ubuntu-latest]
+        python-version: [3.7]
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        onda update -q conda
+        conda config --add channels conda-forge
+        conda config --set channel_priority strict
+        conda env create --file="ci/travis/37-mc.yaml"
+    - name: Lint with black
+      run: |
+        black --check .
+    - name: Test with pytest
+      run: |
+        pytest -v --cov-config .coveragerc --cov=momepy tests
