+++ .github/workflows/momepy_CI.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       max-parallel: 9
       matrix:
-        os: [macos-latest, windows-latest, ubuntu-latest]
+        os: [ubuntu-latest]
         python-version: [3.7]
 
     steps:
@@ -20,6 +20,9 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
+        export PATH="$HOME/miniconda/bin:$PATH"
+        hash -r
+        onda config --set always_yes yes --set changeps1 no
         conda update -q conda
         conda config --add channels conda-forge
         conda config --set channel_priority strict
