### 2019-11-14 b4f36a1
get miniconda
