### 2019-11-14 ea2c618
enable dev, remove travis
