+++ ci/dev.yaml
@@ -10,8 +10,11 @@ dependencies:
   - pytest
   - pytest-cov
   - codecov
-  - mapclassify
+  - pysal
   - pip
   - pip:
-    - inequality
     - black
+    - nose
+    - git+https://github.com/geopandas/geopandas.git
+    - git+https://github.com/pysal/libpysal.git
+    - git+https://github.com/networkx/networkx.git
