+++ ci/travis/dev.yaml
@@ -1,17 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
-  - geopandas
-  - libpysal
-  - networkx
-  - tqdm
-  - pytest
-  - pytest-cov
-  - codecov
-  - pysal
-  - pip
-  - pip:
-    - black
-    - nose
