+++ ci/travis/37-pysal.yaml
@@ -1,16 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
-  - geopandas
-  - libpysal
-  - networkx
-  - tqdm
-  - pytest
-  - pytest-cov
-  - codecov
-  - pysal
-  - pip
-  - pip:
-    - black
