+++ .travis.yml
@@ -1,40 +1,0 @@
-language: python
-
-sudo: false
-
-matrix:
-  include:
-    - env: ENV_FILE="ci/travis/37-pysal.yaml"
-    - env: ENV_FILE="ci/travis/37-mc.yaml" STYLE=true
-    - env: ENV_FILE="ci/travis/dev.yaml" DEV=true
-
-install:
-  - sudo apt-get update
-  # We do this conditionally because it saves us some downloading if the
-  # version is the same.
-  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  # Useful for debugging any issues with conda
-  - conda config --add channels conda-forge
-  - conda config --set channel_priority strict
-  - conda info -a
-
-  - conda env create --file="${ENV_FILE}"
-  - source activate test
-  - if [ "$DEV" ]; then pip install git+https://github.com/geopandas/geopandas.git; fi
-  - if [ "$DEV" ]; then pip install git+https://github.com/pysal/libpysal.git; fi
-  - if [ "$DEV" ]; then pip install git+https://github.com/networkx/networkx.git; fi
-
-  - python setup.py install
-  - conda list
-
-script:
-  - pytest -v --cov-config .coveragerc --cov=momepy tests
-  - if [ "$STYLE" ]; then black --check .; fi
-
-after_success:
-  - codecov
