+++ .github/workflows/momepy_CI.yml
@@ -10,15 +10,10 @@ jobs:
       max-parallel: 9
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.7]
         envs: [ci/37-mc.yaml, ci/37-pysal.yaml, ci/dev.yaml]
 
     steps:
     - uses: actions/checkout@v1
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v1
-      with:
-        python-version: ${{ matrix.python-version }}
     - name: Build environment
       run: |
         pushd .
@@ -39,6 +34,9 @@ jobs:
         conda env create --file="${{ matrix.envs }}"
         conda activate test
         python setup.py install
+    - name: Conda list
+      run: |
+        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
         conda list
     - name: Black
       run: |
@@ -47,4 +45,9 @@ jobs:
     - name: Run tests
       run: |
         eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
-        pytest -v tests
+        pytest --cov=./ --cov-report=xml -v tests
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v1.0.3
+      with:
+        token: ${{secrets.CODECOV_TOKEN}}
+        file: ./coverage.xml
