+++ .github/workflows/macos_CI.yml
@@ -1,4 +1,4 @@
-name: momepy CI
+name: Linux
 
 on: [push]
 
@@ -7,10 +7,10 @@ jobs:
 
     runs-on: ${{ matrix.os }}
     strategy:
-      max-parallel: 9
+      max-parallel: 3
       matrix:
-        os: [ubuntu-latest]
-        envs: [ci/37-mc.yaml, ci/37-pysal.yaml, ci/dev.yaml]
+        os: [macOS-latest]
+        envs: [ci/37-mc.yaml]
 
     steps:
     - uses: actions/checkout@v1
@@ -18,7 +18,7 @@ jobs:
       run: |
         pushd .
         cd $HOME
-        MINICONDA=Miniconda3-latest-Linux-x86_64.sh
+        MINICONDA=Miniconda3-latest-MacOSX-x86_64.sh
         MINICONDA_HOME=$HOME/miniconda
         MINICONDA_MD5=$(curl -s https://repo.continuum.io/miniconda/ | grep -A3 $MINICONDA | sed -n '4p' | sed -n 's/ *<td>\(.*\)<\/td> */\1/p')
         wget -q https://repo.continuum.io/miniconda/$MINICONDA
@@ -38,16 +38,7 @@ jobs:
       run: |
         eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
         conda list
-    - name: Black
-      run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
-        black --check .
     - name: Run tests
       run: |
         eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
-        pytest --cov=./ --cov-report=xml -v tests
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v1.0.3
-      with:
-        token: ${{secrets.CODECOV_TOKEN}}
-        file: ./coverage.xml
+        pytest -v tests
