+++ ci/dev.yaml
@@ -4,7 +4,6 @@ channels:
 dependencies:
   - python=3.7
   - geopandas
-  - libpysal
   - networkx
   - tqdm
   - pytest
