+++ .github/workflows/linux_CI.yml
@@ -1,4 +1,4 @@
-name: Linux
+name: Linux CI
 
 on: [push]
 
@@ -18,12 +18,7 @@ jobs:
       run: |
         pushd .
         cd $HOME
-        MINICONDA=Miniconda3-latest-Linux-x86_64.sh
-        MINICONDA_HOME=$HOME/miniconda
-        MINICONDA_MD5=$(curl -s https://repo.continuum.io/miniconda/ | grep -A3 $MINICONDA | sed -n '4p' | sed -n 's/ *<td>\(.*\)<\/td> */\1/p')
-        wget -q https://repo.continuum.io/miniconda/$MINICONDA
-        bash $MINICONDA -b -p $MINICONDA_HOME
-        eval "$($MINICONDA_HOME/bin/conda shell.bash hook)"
+        eval "$($CONDA/bin/conda shell.bash hook)"
         conda activate
 
         conda config --set always_yes yes --set changeps1 no
