### 2019-11-14 a62f978
fix conda path
