+++ .github/workflows/linux_CI.yml
@@ -31,15 +31,15 @@ jobs:
         python setup.py install
     - name: Conda list
       run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
+        eval "$($CONDA/bin/conda shell.bash hook)" && conda activate test
         conda list
     - name: Black
       run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
+        eval "$($CONDA/bin/conda shell.bash hook)" && conda activate test
         black --check .
     - name: Run tests
       run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
+        eval "$($CONDA/bin/conda shell.bash hook)" && conda activate test
         pytest --cov=./ --cov-report=xml -v tests
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1.0.3
