### 2019-11-14 5a85d0d
use local conda on mac
