+++ .github/workflows/macos_CI.yml
@@ -18,12 +18,7 @@ jobs:
       run: |
         pushd .
         cd $HOME
-        MINICONDA=Miniconda3-latest-MacOSX-x86_64.sh
-        MINICONDA_HOME=$HOME/miniconda
-        MINICONDA_MD5=$(curl -s https://repo.continuum.io/miniconda/ | grep -A3 $MINICONDA | sed -n '4p' | sed -n 's/ *<td>\(.*\)<\/td> */\1/p')
-        wget -q https://repo.continuum.io/miniconda/$MINICONDA
-        bash $MINICONDA -b -p $MINICONDA_HOME
-        eval "$($MINICONDA_HOME/bin/conda shell.bash hook)"
+        eval "$($CONDA/bin/conda shell.bash hook)"
         conda activate
 
         conda config --set always_yes yes --set changeps1 no
@@ -36,9 +31,9 @@ jobs:
         python setup.py install
     - name: Conda list
       run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
+        eval "$($CONDA/bin/conda shell.bash hook)" && conda activate test
         conda list
     - name: Run tests
       run: |
-        eval "$($HOME/miniconda/bin/conda shell.bash hook)" && conda activate test
+        eval "$($CONDA/conda shell.bash hook)" && conda activate test
         pytest -v tests
