+++ ci/travis/37-mc.yaml
@@ -2,7 +2,7 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python
+  - python=3.6
   - geopandas
   - libpysal
   - networkx
