+++ ci/travis/tst.yaml
@@ -1,0 +1,6 @@
+name: test
+channels:
+  - conda-forge
+dependencies:
+  - python=3.7
+  - mapclassify
