+++ ci/travis/tst.yaml
@@ -2,5 +2,4 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python=3.7
   - mapclassify
