+++ ci/travis/latest-mc.yaml
@@ -2,7 +2,7 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python=3.7
+  - python
   - geopandas
   - libpysal
   - networkx
@@ -10,7 +10,8 @@ dependencies:
   - pytest
   - pytest-cov
   - codecov
-  - pysal
+  - mapclassify
   - pip
   - pip:
+    - inequality
     - black
