+++ ci/travis/tst.yaml
@@ -1,6 +1,0 @@
-name: test
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
-  - scikit-learn
