### 2019-11-14 f29776a
skipif mapclassify version
