### 2019-11-14 86cbe26
skipif pysal, mc
