+++ tests/test_diversity.py
@@ -2,9 +2,17 @@ import geopandas as gpd
 import momepy as mm
 import numpy as np
 import pytest
-import mapclassify
 from momepy import sw_high
 from pytest import approx
+
+from distutils.version import LooseVersion
+
+try:
+    import mapclassify
+    MC_21 = str(mapclassify.__version__) <= LooseVersion("2.1.0")
+except ImportError:
+    import pysal
+    MC_21 = str(pysal.__version__) <= LooseVersion("2.1.0")
 
 
 class TestDiversity:
@@ -45,7 +53,7 @@ class TestDiversity:
         ).series
         assert zeros[0] == 0
 
-    @pytest.mark.skipif(mapclassify.__version__ < (2, 1), reason="mapclassify < 2.1")
+    @pytest.mark.skipif(MC_21, reason="requires mapclassify < 2.1")
     def test_Simpson(self):
         ht_sw = mm.Simpson(self.df_tessellation, "area", self.sw, "uID").series
         assert ht_sw[0] == 0.385
