### 2019-11-14 ea4950c
CI: Use python 3.8, allow failures in DEV (#119)
* CI: Use python 3.8, allow failures in DEV

* skipif mapclassify version
