+++ ci/travis/dev.yaml
@@ -2,7 +2,7 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python=3.7
+  - python=3.8
   - geopandas
   - libpysal
   - networkx
@@ -10,8 +10,9 @@ dependencies:
   - pytest
   - pytest-cov
   - codecov
-  - pysal
+  - mapclassify
   - pip
   - pip:
     - black
     - nose
+    - inequality
