### 2019-11-14 d10d4d1
Create .pre-commit-config.yaml
