### 2019-11-16 #1807 @drvinceknight
@martinfleis it looks like https://doi.org/10.1007/978-3-642-03647-7_11 is a DOI for one of the papers you cite, could you add that please?

Once you've done that if you could make a tagged release and archive, and report the version number and archive DOI here please.
