+++ paper/paper.bib
@@ -83,7 +83,7 @@ pages = {11--15},
 address = {Pasadena, CA USA},
 }
 
-@software{kelsey_jordahl_2019_3463125,
+@software{kelsey_jordahl_2019_3483425,
   author       = {Kelsey Jordahl and
                   Joris Van den Bossche and
                   Jacob Wasserman and
@@ -114,13 +114,13 @@ address = {Pasadena, CA USA},
                   Filipe and
                   Chris Holdgraf and
                   Adam Greenhall},
-  title        = {geopandas/geopandas: v0.6.0},
-  month        = sep,
+  title        = {geopandas/geopandas: v0.6.1},
+  month        = oct,
   year         = 2019,
   publisher    = {Zenodo},
-  version      = {v0.6.0},
-  doi          = {10.5281/zenodo.3463125},
-  url          = {https://doi.org/10.5281/zenodo.3463125}
+  version      = {v0.6.1},
+  doi          = {10.5281/zenodo.3483425},
+  url          = {https://doi.org/10.5281/zenodo.3483425}
 }
 
 @book{Kropf:2017tw,
@@ -147,15 +147,24 @@ month = dec
 }
 
 
-@article{pysal2007,
-author = {Rey, Sergio J and Anselin, Luc},
-title = {{PySAL: A Python Library of Spatial Analytical Methods}},
-journal = {The Review of Regional Studies},
-year = {2007},
-volume = {37},
-number = {1},
-pages = {5--27},
+@Inbook{Rey2010,
+author="Rey, Sergio J.
+and Anselin, Luc",
+editor="Fischer, Manfred M.
+and Getis, Arthur",
+title="PySAL: A Python Library of Spatial Analytical Methods",
+bookTitle="Handbook of Applied Spatial Analysis: Software Tools, Methods and Applications",
+year="2010",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="175--193",
+abstract="This chapter describes PySAL, an open source library for spatial analysis written in the object oriented language Python. PySAL grew out of the software development activities that were part of the Center for Spatially Integrated Social Sciences Tools Project (Goodchild et al. 2000). This National Science Foundation infrastructure project had as its goals to facilitate dissemination of spatial analysis software to social sciences, to develop a library of spatial data analysis modules, to develop prototypes implementing state of the art methods, and to initiate and nurture a community of open source developers.",
+isbn="978-3-642-03647-7",
+doi="10.1007/978-3-642-03647-7_11",
+url="https://doi.org/10.1007/978-3-642-03647-7_11"
 }
+
+
 
 @article{Sevtsuk:2012ixa,
 author = {Sevtsuk, Andres and Mekonnen, Michael},
