+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-Urban morphology is the area of urban studies that explores the physical form of cities in space and the way it changes in time in relation to the agents of such change [@Kropf:2017tw]. The discipline is based on the analysis of space, traditionally mostly visual and qualitative [@2015arXiv150604875D]; its objects are the fundamental elements of urban form (building, plot, street) [@Moudon:1997wk] as well as a range of analytical constructs such as axial maps [@ArizaVillaverde:2013bt] or proximity bands [@Araldi:2019kv]. The increased availability of morphological data and computational power have led in time to more emphasis on quantitative forms of analysis, and the emergence of Urban Morphometrics (UMM) [@Dibble:2017bz]: this approach describes urban form via the systematic and comprehensive measurement of its morphological characters. Since UMM analysis is addressed both in-depth and at large scale, it is grounded on the intensive use of GIS software (proprietary ArcGIS, MapInfo, open-source QGIS) either through built-in processing tools or specific plugins like Urban Network Analysis [@Sevtsuk:2012ixa] or Place Syntax Tool [@Stahle469861]. However, essential functions to conduct such measurements on specific urban morphometric characters or processes such as morphological tessellation, are not always available: current plugins offer only a limited number of functionalities as they are mainly application or case-specific. We are hereby proposing ``momepy``, a Python toolkit which aims to overcome such limitations by enabling a systematic in-depth analysis of urban form, to comprehensively include a wide range of measurable characters, with a prospect of expanding future development due to its open-source nature and independence on proprietary software or operating systems. The development of momepy is timely, as the role of measurable characters is vital to recognize form-based patterns and establish descriptive and analytical frameworks of human settlements, in the “age of urbanization”.
+Urban morphology is the area of urban studies that explores the physical form of cities in space and the way it changes in time in relation to the agents of such change [@Kropf:2017tw]. The discipline is based on the analysis of space, traditionally mostly visual and qualitative [@2015arXiv150604875D]; its objects are the fundamental elements of urban form (building, plot, street) [@Moudon:1997wk] as well as a range of analytical constructs such as axial maps [@ArizaVillaverde:2013bt] or proximity bands [@Araldi:2019kv]. The increased availability of morphological data and computational power have led in time to more emphasis on quantitative forms of analysis, and the emergence of Urban Morphometrics (UMM) [@Dibble:2017bz]: this approach describes urban form via the systematic and comprehensive measurement of its morphological characters. Since UMM analysis is addressed both in-depth and at large scale, it is grounded on the intensive use of GIS software (proprietary ArcGIS, MapInfo, open-source QGIS) either through built-in processing tools or specific plugins like Urban Network Analysis [@Sevtsuk:2012ixa] or Place Syntax Tool [@Stahle469861]. However, essential functions to conduct such measurements on specific urban morphometric characters or processes such as morphological tessellation, are not always available: current plugins offer only a limited number of functionalities as they are mainly application or case-specific. This paper is hereby proposing ``momepy``, a Python toolkit which aims to overcome such limitations by enabling a systematic in-depth analysis of urban form, to comprehensively include a wide range of measurable characters, with a prospect of expanding future development due to its open-source nature and independence on proprietary software or operating systems. The development of momepy is timely, as the role of measurable characters is vital to recognize form-based patterns and establish descriptive and analytical frameworks of human settlements, in the “age of urbanization”.
 
 The morphometric characters implemented in momepy allow to conduct a cross-scale analysis of urban form, spanning from the level of individual buildings and plots to the macroscopic metropolitan regions. Furthermore, ``momepy`` is designed to be more flexible than other toolkits as its functions are generally not restricted to specific morphological elements but to geometry types only and as such, can be used in various analytical models.
 
