### 2019-11-16 #1807 @martinfleis
@drvinceknight Thank you! This is another paper, but I exchanged them as they both talk about the same and this one has a DOI.

I just made a release v.0.1.0 (https://github.com/martinfleis/momepy/releases/tag/v0.1.0)
Zenodo DOI for it is 10.5281/zenodo.3544609 (https://zenodo.org/record/3544609#.XdB3ui9OzOQ)
