### 2019-11-20 #1807 @martinfleis
Hi @drvinceknight, do you think we'll be able to finish this review this week? I need to understand whether I can reference it elsewhere (and hence have an active doi) or not yet. Thanks!
