### 2019-11-20 #1807 @drvinceknight
> Hi @drvinceknight, do you think we'll be able to finish this review this week? I need to understand whether I can reference it elsewhere (and hence have an active doi) or not yet. Thanks!

Yup, we should definitely be done by end of the week. Could you modify the archive metadata to match the paper. Specifically the title needs to be: `momepy: Urban Morphology Measuring Toolkit`
