### 2019-11-20 #1807 @drvinceknight
@openjournals/joss-eics this paper is ready to be accepted.

Thanks again to @geraintpalmer and @martibosch for your time and effort reviewing this work. :+1:
